+++ atlas/modules/transformer/ordering/resource.py
@@ -1,6 +1,5 @@
-import re
-
-from atlas.modules import constants, utils
+from atlas.modules import constants, exceptions, utils
+from atlas.modules.helpers import open_api
 from atlas.modules.transformer.ordering.base import Node, DAG
 
 
@@ -9,6 +8,54 @@ class Reference:
     def __init__(self, key, config):
         self.name = key
         self.config = config
+
+        self.connected_refs = set()
+        self.connected_resources = set()
+
+        self.primary_resource = None
+
+    def add_connected_ref(self, value):
+        if value:
+            self.connected_refs.add(value)
+
+    def add_connected_resource(self, value):
+        if value:
+            self.connected_resources.add(value)
+
+    def resolve_primary_resource(self, candidates: set):
+        candidates = candidates or self.connected_resources
+
+        # Only one candidate is there, so we have clear primary resource
+        if len(candidates) == 1:
+            self.primary_resource = self.connected_resources.pop()
+        elif not candidates:
+            # There are no candidates, so just name resource after reference
+            self.primary_resource = self.name
+        else:
+            # We cannot determine the best candidate
+            raise exceptions.ResourcesException(f"Could not determine primary resource for {self.name}")
+
+    def get_connections(self):
+
+        primary_resource_fields = set()
+
+        # Look through properties
+        for name, config in self.config.get(constants.PROPERTIES, {}).items():
+            field = open_api.ReferenceField(name, config)
+            self.add_connected_ref(field.ref)
+            self.add_connected_resource(field.resource)
+
+            if field.can_contain_primary_resource:
+                primary_resource_fields.add(field.resource)
+
+        # Now look through Additional properties to see if there is any ref there
+        additional_properties = self.config.get(constants.ADDITIONAL_PROPERTIES, {})
+
+        if additional_properties and isinstance(additional_properties, dict):
+            self.add_connected_ref(additional_properties.get(constants.REF))
+            self.add_connected_resource(additional_properties.get(constants.RESOURCE))
+
+        self.resolve_primary_resource(primary_resource_fields)
 
 
 class Resource(Node):
@@ -37,121 +84,93 @@ class ResourceGraph(DAG):
     def __init__(self, references: dict, specs=None):
         super(ResourceGraph, self).__init__()
         self.resources = dict()
-        self.references = [Reference(key.lower(), config) for key, config in references.items()]
+        self.references = {key.lower(): Reference(key.lower(), config) for key, config in references.items()}
+        # self.references = [Reference(key.lower(), config) for key, config in references.items()]
         self.specs = specs or {}
 
-    @staticmethod
-    def convert_ref_name_to_resource(ref_name: str) -> str:
-        return "".join([x.lower() for x in re.sub("-", "_", ref_name).split("_")])
+    def get_associated_resource_for_ref(self, ref_name: str) -> str:
+        return self.references[ref_name.lower()].primary_resource
 
     def add_ref_edge(self, ref, dependent_key):
+        """
+        Add resource edge by extracting resource from ref
+        """
         if ref:
             source_key = utils.get_ref_name(ref)
-            source_resource = self.convert_ref_name_to_resource(source_key)
+            source_resource = self.get_associated_resource_for_ref(source_key)
             # Note the edge - Source is required for resource
             self.add_edge(source_resource, dependent_key)
-
-    @staticmethod
-    def process_resource_for_refs(resource_config: dict) -> (set, set):
-        """
-        Process a single resource and return set of all references from it
-        """
-        refs = set()
-        resources = set()
-
-        # ########## --  Ignore ALL OF, since adding these relations gave us several cycles -- ####
-
-        # # Check if there is reference
-        # # This is needed when this is called recursively via All of mechanism
-        # refs.add(resource_config.get(constants.REF))
-
-        # # Recursively process ALL OF sub-schemas
-        # for schema in resource_config.get(constants.ALL_OF, []):
-        #     refs.update(self.process_resource_for_refs(schema))
-
-        # ####### -- Rest of code starts here ------- ###################
-
-        # Look through properties
-        for config in resource_config.get(constants.PROPERTIES, {}).values():
-            _type = config.get(constants.TYPE)
-            if _type == constants.ARRAY:
-                config = config.get(constants.ITEMS, {})
-            ref = config.get(constants.REF)
-            refs.add(ref)
-            resource = config.get(constants.RESOURCE)
-            resources.add(resource)
-
-        # Now look through Additional properties to see if there is any ref there
-        additional_properties = resource_config.get(constants.ADDITIONAL_PROPERTIES, {})
-
-        if additional_properties and isinstance(additional_properties, dict):
-            refs.add(additional_properties.get(constants.REF))
-            resources.add(additional_properties.get(constants.RESOURCE))
-
-        return refs, resources
 
     def construct_graph(self):
 
         # Add all nodes first
-        for reference in self.references:
-            resource_name = self.convert_ref_name_to_resource(reference.name)
-            self.resources[resource_name] = reference.config
+        for ref_name, reference in self.references.items():
+            reference.get_connections()
+            resource_name = self.get_associated_resource_for_ref(ref_name)
+            self.resources[resource_name] = reference
             self.add_node(resource_name)
 
         # Now add edges between the nodes
-        for resource_key, resource_config in self.resources.items():
-            refs, resources = self.process_resource_for_refs(resource_config)
-            for ref in refs:
+        for resource_key, reference in self.resources.items():
+            for ref in reference.connected_refs:
                 self.add_ref_edge(ref, resource_key)
-            for resource in resources:
+            for resource in reference.connected_resources:
                 if resource:
                     self.add_edge(resource, resource_key)
 
     @staticmethod
-    def get_ref_name(config):
-        schema = config.get(constants.SCHEMA, {})
-
-        # Search in simple direct ref or array ref
-        ref = schema.get(constants.REF)
-
-        if schema and not ref:
-            items = schema.get(constants.ITEMS, {})
-            if items and isinstance(items, dict):
-                ref = items.get(constants.REF)
-
-        return utils.get_ref_name(ref).lower() if ref else None
+    def get_schema_refs(config):
+        schema = open_api.Schema(config.get(constants.SCHEMA, {}))
+        return [utils.get_ref_name(ref).lower() for ref in schema.get_all_refs()]
 
     def parse_paths(self, interfaces):
 
         for operation in interfaces:
             op_id = operation.func_name
             ref_graph = {}
-            for response in operation.responses.values():
-                if operation.method in {constants.DELETE, constants.PATCH, constants.PUT}:
-                    continue
-                ref = self.get_ref_name(response)
-                if ref:
-                    ref_graph[ref] = "producer"
-            for parameter in operation.parameters.values():
-                # Try to find if it is Path Params Resource
-                # This should over-write any previous value
-
-                parameter_ref = parameter.get(constants.REF)
-                if parameter_ref:
-                    parameter = utils.resolve_reference(self.specs, parameter_ref)
-
-                resource = parameter.get(constants.RESOURCE)
-                if resource:
-                    # This time, it is Path Params, so we are sure that is is consumer
-                    ref_graph[resource] = "consumer"
-
-                ref = self.get_ref_name(parameter)
-                # If it is already claimed as producer/consumer, then respect that
-                if ref and ref not in ref_graph:
-                    ref_graph[ref] = "consumer"
+            self.parse_responses(operation, ref_graph)
+            self.parse_request_parameters(operation, ref_graph)
 
             for ref, ref_op in ref_graph.items():
-                resource_node = self.nodes.get(self.convert_ref_name_to_resource(ref))
+                resource_node = self.nodes.get(self.get_associated_resource_for_ref(ref))
                 ref_op = "add_consumer" if ref_op == "consumer" else "add_producer"
                 if resource_node:
                     getattr(resource_node, ref_op)(op_id)
+
+    def parse_responses(self, operation, ref_graph):
+        """
+        Parse Responses for a single operation and mark References w.r.t to Operation
+        """
+
+        for response in operation.responses.values():
+            if operation.method in {constants.DELETE, constants.PATCH, constants.PUT}:
+                continue
+            response_refs = self.get_schema_refs(response)
+            for ref in response_refs:
+                ref_graph[ref] = "producer"
+
+    def parse_request_parameters(self, operation, ref_graph):
+        """
+        Parse Request Parameters for a single operation and mark References w.r.t to Operation
+        """
+
+        for parameter in operation.parameters.values():
+            # Try to find if it is Path Params Resource
+            # This should over-write any previous value
+
+            parameter_ref = parameter.get(constants.REF)
+            if parameter_ref:
+                parameter = utils.resolve_reference(self.specs, parameter_ref)
+
+            resource = parameter.get(constants.RESOURCE)
+            if resource:
+                # This time, it is Path Params, so we are sure that is is consumer
+                ref_graph[resource] = "consumer"
+
+            if parameter.get(constants.TYPE) == constants.BODY_PARAM:
+                request_refs = self.get_schema_refs(parameter)
+
+                # If it is already claimed as producer/consumer, then respect that
+                for ref in request_refs:
+                    if ref not in ref_graph:
+                        ref_graph[ref] = "consumer"