+++ atlas/settings.py
@@ -14,20 +14,32 @@ class Settings:
     OUTPUT_FOLDER = "build"
 
     SWAGGER_FILE = "swagger.yaml"
-    LOCUST_FILE = "locust.py"
-    ARTILLERY_FILE = "processor.js"
-    ARTILLERY_YAML = "artillery.yaml"
-
-    LOCUST_HOOK_FILE = "hooks.py"
-    ARTILLERY_HOOK_FILE = "hooks.js"
     MAPPING_FILE = "resource_mapping.yaml"
     RES_MAPPING_HOOKS_FILE = "resource_hooks.py"
     PROFILES_FILE = "profiles.yaml"
     RESOURCES_FOLDER = "resources"
     DIST_FOLDER = "dist"
 
+    LOCUST_FILE = "locust.py"
+    LOCUST_HOOK_FILE = "hooks.py"
+
     # Generated Output files for Artillery
     # Do not change these names, as they are imported as it is in JS
     ARTILLERY_LIB_FOLDER = "libs"
     ARTILLERY_PROFILES = "profiles.js"
     ARTILLERY_RESOURCES = "resources.js"
+    ARTILLERY_FILE = "processor.js"
+    ARTILLERY_YAML = "artillery.yaml"
+    ARTILLERY_HOOK_FILE = "hooks.js"
+
+    # ### Resource Auto-detection settings
+    # These suffixes if present determine whether the URL parameter is resource or not
+    URL_PARAM_RESOURCE_SUFFIX = {"_id", "Id", "_slug", "Slug", "pk"}
+
+    # These are direct names for Path parameters
+    PATH_PARAM_RESOURCES = {"id", "slug", "pk"}
+
+    # These are field names which if present in references are marked as resource with Reference name
+    REFERENCE_FIELD_RESOURCES = {"id", "slug", "pk"}
+
+    # #### End Resource Auto-detection constant settings